MNT: move make_graph method to LightController
